requirements.txt
requests
yfinance
pandas
jina
openai
numpy
logging
python-dotenv